requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-phonenumber-field==7.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==2.0.0
pandas==2.2.2
phonenumbers==8.13.39
pillow==10.3.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
